requirements.txt
numpy
scipy
matplotlib
astropy
nbsphinx
